Registry: Try to handle circular references
An ASN type may refer to itself, this happens for example in NR-RRC-Definitions for the type RRCReconfigurationComplete

Handle it by adding the translator to the register before the recursive call.
In order to store the translator before it exists (it will exist only after the recursive call ends) a special wrapper is used.